Update replaceNullWithNa
BUG: @rolson-cor noticed an issue with outputting the HTML report while setting up the GPP repository. The issue arose in the drugexp and drugevents datasets. The issue had to do with looping through the "expectedValues" in the codebook and looking to replace the vector-within-a-column for the expectedValues. It was looking for a size of 1 but found a size of 8.

CAUSE: As of now - the true cause is unknown. It's unclear why this is proccing now and hasn't for any of the other registries. However - steps were taken to remedy.

FIX: Updated the replaceNullWithNa code to collapse any vector into a single character string separated by commas to prevent this issue.

- Updated replaceNullWithNa to collapse the vector-valued-cell into a text string with the values separated by commas